+++ src/utils/elementChangers.js
@@ -1,7 +1,6 @@
 import { buttonTabPreValue } from "./defaults/defaultVariables"
 
 export function showHideStatusBar(showHideState,appGlobalArgs) {
-  let eEWidgetElem = document.getElementsByClassName('module-timeEEWidget')[0]
   if(showHideState) {
     appGlobalArgs.elem_statusBar.classList.remove('statusBar--hidden')
   }else {