+++ src/utils/modules/moduleGlobalUtils.js
@@ -1,5 +1,5 @@
-import { importPlatformsData } from "../module/platforms/platforms.import.js"
-import { importFiltersData } from "../module/filters/filters.import.js"
+import { importPlatformsData } from "../../module/platforms/platforms.import.js"
+import { importFiltersData } from "../../module/filters/filters.import.js"
 
 export async function reimportModuleStorageData(moduleCategoryName){
     let importNewData = ''